+++ rumqttd/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - v4 config is optional, user can specify v4 and/or v5 config
 - websocket feature is enabled by default
 - console configuration is optional
+- rustls client auth is featured gated behind "verify-client-cert" ( disabled by default ). 
 
 ### Deprecated
 - "websockets" feature is removed in favour of "websocket"